requirements.txt
asgiref==3.4.1
Django==4.0.1
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
uWSGI==2.0.20
